update LfaAllowed logic to prevent races until it can be reworked for per-execution ID
